requirements.txt
streamlit
openai
langchain
chromadb
sentence-transformers
python-dotenv
uvicorn
fastapi
PyPDF2
tqdm
pytest
